% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crepuscule.R
\docType{package}
\name{crepuscule}
\alias{crepuscule}
\alias{crepuscule-package}
\title{crepuscule: finds sun events in a day}
\description{
get_sun_events is the main function transforming a data.frame with latitude,
longitude and dateTime to calculate nautical and civil twilight as well as
sunrise and sunset
}

