% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_event_function.R
\name{get_sun_events}
\alias{get_sun_events}
\title{Get sun events from latitude, longitude and date}
\usage{
get_sun_events(data)
}
\arguments{
\item{data}{a data.frame with column longitude, latitude and dateTime}
}
\value{
a data.frame containing both data and new columns with dawn, dusk, sunrise, sunset ...
}
\description{
\code{get_sun_events} calculates nautical and civil dawn and dusk as well as sunrise and sunset, daylength, nautical length and civil length.
}
\examples{
get_sun_events(data)
}

