% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_day_night.R
\name{get_time2sunevent}
\alias{get_time2sunevent}
\title{Get day, night and time to sunrise or sunset}
\usage{
get_time2sunevent(time, sunrise, sunset)
}
\description{
Get day, night and time to sunrise or sunset
}

